v0.2 - Fixed certain parts showing in stock when they weren't
Some parts return data from the API, including prices and stock information, but aren't actually available in JLC's library. An example is C242505 - an ME2-CO sensor which drew my suspicions because it's a TH part, but it showed up as in stock. These parts have a bunch of null values for things like componentName and describe fields. Updated code marks these as N/A.